Update responsive breakpoints and custom styles
Refine website responsive design by adjusting CSS and SCSS breakpoints. Modify heights and min-heights in the homepage slider and other pages, and tweak existing styles in WooCommerce and the global sections. Transform properties are applied to improve layout on devices with a max-width of 390px. Changes aim to enhance user interface and overall user experience.